Ajustes em DTOs e serviços para suportar valores nulos
As alterações no arquivo `PedidoDto.cs` incluem a mudança do tipo de `ClienteId` de `int` para `int?`, permitindo que ele seja nulo. Além disso, `Itens` e `ValorTotal` também foram alterados para permitir valores nulos, e `ValorTotal` agora é uma propriedade que pode ser definida em vez de ser calculada diretamente a partir dos itens.

No arquivo `PedidoItemDto.cs`, várias propriedades foram alteradas para serem do tipo `int?` ou `decimal?`, permitindo que sejam nulas. A propriedade `Subtotal` foi alterada para ser uma propriedade que pode ser definida em vez de ser calculada diretamente.

No `PedidoItemMapper`, foram adicionadas as propriedades `PedidoId` e `Subtotal` ao mapeamento, e o método `ToModel` foi atualizado para lidar com as novas propriedades nulas. O `PedidoMapper` foi modificado para lidar com as novas propriedades nulas e agora chama `AtualizarValorTotal` após a criação do pedido.

No `PedidoItemService`, um novo repositório `IProdutoRepository` foi adicionado ao construtor, e várias validações foram atualizadas para lidar com as novas propriedades nulas. O método `Adicionar` agora lança exceções mais específicas em caso de erro, e o método `Atualizar` foi modificado para verificar se o `dto` é válido.

O `IClienteRepository` foi alterado para garantir que o método `ObterPorId` sempre retorne um `Cliente` em vez de um `Cliente?`. No `Pedido`, a propriedade `ValorTotal` agora é calculada a partir de um campo privado `_valorTotal`, e um novo método `AtualizarValorTotal` foi adicionado.

No `PedidoItem`, as propriedades `PrecoUnitario` e `Subtotal` foram alteradas para serem calculadas a partir de campos privados, e um novo método `AtualizarValores` foi adicionado.

Os testes em `PedidoItemServiceTests` e `PedidoServiceTests` foram atualizados para refletir as mudanças nas propriedades e métodos, incluindo a adição de mocks para o novo repositório de produtos e ajustes nas chamadas de métodos para lidar com as novas definições de propriedades.

+++ CRM.Tests/Servicos/PedidoServiceTests.cs
@@ -38,10 +38,10 @@ public class PedidoServiceTests
             }
         };
 
-        _clienteRepositoryMock.Setup(r => r.ObterPorId(pedidoDto.ClienteId))
+        _clienteRepositoryMock.Setup(r => r.ObterPorId((int)pedidoDto.ClienteId))
             .ReturnsAsync(new Cliente
             {
-                Id = pedidoDto.ClienteId,
+                Id = (int)pedidoDto.ClienteId,
                 Nome = "João",
                 Email = "joao@email.com",
                 Endereco = "endereço",
@@ -59,7 +59,7 @@ public class PedidoServiceTests
             .ReturnsAsync(new Pedido
             {
                 Id = 1,
-                ClienteId = pedidoDto.ClienteId,
+                ClienteId = (int)pedidoDto.ClienteId,
                 Itens = new List<PedidoItem>
                 {
                     new PedidoItem { ProdutoId = 1, Quantidade = 2 }
